Re-introduce remove member
